requirements.txt
python-dotenv
fastapi
uvicorn
requests
langchain-openai
langchain